# docs/architecture/api/openapi.yaml
openapi: 3.1.0
info:
  title: CWMS Construction Work Management System API
  version: 1.0.0
  summary: Implementation-agnostic HTTP API for CWMS v1.0 web application
  description: |
    Architecture-level OpenAPI 3.1 specification for CWMS Version 1.0.
    Not tied to any framework. Traceability to Functional Specification is via
    `x-fs-refs` on operations and the companion API-CATALOGUE.md.
    Money amounts are decimal strings. Dates are ISO-8601.
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Version 1 API base path
...
security:
  - sessionAuth: []
  - bearerAuth: []
paths:
...
  /works/{workId}/schedule-activities:
    parameters:
      - $ref: '#/components/parameters/WorkId'
    get:
      tags: [Schedule]
      operationId: listScheduleActivities
      x-fs-refs: [F-SCH-01]
      x-prd-refs: [PRD-SCH]
      responses:
        '200':
          description: Activities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleActivityList'
    post:
      tags: [Schedule]
      operationId: createScheduleActivity
      x-fs-refs: [F-SCH-01]
      x-br-refs: [BR-STAT-04, BR-STAT-05]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleActivityCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleActivity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /schedule-activities/{activityId}:
    parameters:
      - $ref: '#/components/parameters/ActivityId'
    patch:
      tags: [Schedule]
      operationId: updateScheduleActivity
      x-fs-refs: [F-SCH-03]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleActivityUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleActivity'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Schedule]
      operationId: deleteScheduleActivity
      x-fs-refs: [F-SCH-03]
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: CWMSSESSION
      description: Server session cookie (implementation-defined name)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer/session token alternative
  parameters:
...
    WorkId:
      name: workId
      in: path
      required: true
      schema: { type: string }
...
    ActivityId:
      name: activityId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    Forbidden:
      description: Not authorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    Percent:
      type: string
      pattern: '^\d+(\.\d{1,4})?$'
      example: "18.00"
...
    ProblemDetails:
      type: object
      required: [title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Machine-stable error code
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
              code: { type: string }
...
    ScheduleActivity:
      type: object
      required: [id, workId, activity]
      properties:
        id: { type: string }
        workId: { type: string }
        activity: { type: string }
        startDate: { type: string, format: date, nullable: true }
        finishDate: { type: string, format: date, nullable: true }
        actualStart: { type: string, format: date, nullable: true }
        actualFinish: { type: string, format: date, nullable: true }
        progressPercent: { $ref: '#/components/schemas/Percent' }
    ScheduleActivityCreateRequest:
      type: object
      required: [activity]
      properties:
        activity: { type: string, minLength: 1 }
        startDate: { type: string, format: date }
        finishDate: { type: string, format: date }
        actualStart: { type: string, format: date }
        actualFinish: { type: string, format: date }
        progressPercent: { $ref: '#/components/schemas/Percent' }
    ScheduleActivityUpdateRequest:
      $ref: '#/components/schemas/ScheduleActivityCreateRequest'
    ScheduleActivityList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ScheduleActivity' }
...